+++ core-ng/src/main/java/core/framework/impl/web/bean/QueryParamMapperHelper.java
@@ -2,7 +2,6 @@ package core.framework.impl.web.bean;
 
 import core.framework.impl.web.service.PathParamHelper;
 import core.framework.json.JSON;
-import core.framework.util.Strings;
 import core.framework.web.exception.BadRequestException;
 
 import java.math.BigDecimal;
@@ -46,18 +45,24 @@ final class QueryParamMapperHelper {   // used by generated QueryParamMapper
         return JSON.toEnumValue(enumValue);
     }
 
+    // deserialization helpers
+    public static String toString(String value) {
+        if (value.isEmpty()) return null;
+        return value;
+    }
+
     public static Integer toInt(String value) {
-        if (Strings.isBlank(value)) return null;
+        if (value.isEmpty()) return null;
         return PathParamHelper.toInt(value);
     }
 
     public static Long toLong(String value) {
-        if (Strings.isBlank(value)) return null;
+        if (value.isEmpty()) return null;
         return PathParamHelper.toLong(value);
     }
 
     public static Double toDouble(String value) {
-        if (Strings.isBlank(value)) return null;
+        if (value.isEmpty()) return null;
         try {
             return Double.valueOf(value);
         } catch (NumberFormatException e) {
@@ -66,7 +71,7 @@ final class QueryParamMapperHelper {   // used by generated QueryParamMapper
     }
 
     public static BigDecimal toBigDecimal(String value) {
-        if (Strings.isBlank(value)) return null;
+        if (value.isEmpty()) return null;
         try {
             return new BigDecimal(value);
         } catch (NumberFormatException e) {
@@ -75,7 +80,7 @@ final class QueryParamMapperHelper {   // used by generated QueryParamMapper
     }
 
     public static Boolean toBoolean(String value) {
-        if (Strings.isBlank(value)) return null;
+        if (value.isEmpty()) return null;
         try {
             return Boolean.valueOf(value);
         } catch (NumberFormatException e) {
@@ -84,12 +89,12 @@ final class QueryParamMapperHelper {   // used by generated QueryParamMapper
     }
 
     public static <T extends Enum<?>> T toEnum(String value, Class<T> valueClass) {
-        if (Strings.isBlank(value)) return null;
+        if (value.isEmpty()) return null;
         return PathParamHelper.toEnum(value, valueClass);
     }
 
     public static ZonedDateTime toZonedDateTime(String value) {
-        if (Strings.isBlank(value)) return null;
+        if (value.isEmpty()) return null;
         try {
             return ZonedDateTime.parse(value, DateTimeFormatter.ISO_OFFSET_DATE_TIME);
         } catch (DateTimeParseException e) {
@@ -98,7 +103,7 @@ final class QueryParamMapperHelper {   // used by generated QueryParamMapper
     }
 
     public static LocalDateTime toDateTime(String value) {
-        if (Strings.isBlank(value)) return null;
+        if (value.isEmpty()) return null;
         try {
             return LocalDateTime.parse(value, DateTimeFormatter.ISO_LOCAL_DATE_TIME);
         } catch (DateTimeParseException e) {
@@ -107,7 +112,7 @@ final class QueryParamMapperHelper {   // used by generated QueryParamMapper
     }
 
     public static LocalDate toDate(String value) {
-        if (Strings.isBlank(value)) return null;
+        if (value.isEmpty()) return null;
         try {
             return LocalDate.parse(value, DateTimeFormatter.ISO_LOCAL_DATE);
         } catch (DateTimeParseException e) {